backend/requirements.txt
# Backend - Python FastAPI Microservices

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
elasticsearch==8.11.1

# AI/ML Libraries
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
spacy==3.7.2
transformers==4.36.2
sentence-transformers==2.2.2

# NLP
nltk==3.8.1
textblob==0.17.1

# API & Integration
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1
celery==5.3.4
kombu==5.3.4

# Scheduling
apscheduler==3.10.4

# Security
cryptography==41.0.7
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
reportlab==4.0.7
pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
